Add workflow for deploying static pages
It also adds the most recent site to root. For now the plan is to allow people to visit old versions of the page simply by navigating there and keeping the old pages around in the repository. It is possible in the future to instead of branches and each update to a gh-pages branch but I don't really see a need for a workflow like that on a small project like this.